[revert] use_default_shell_env = True (#10)
### Description

Originally it was unknown to me that we could use flags to directly disable the stamping of scala imports using 
`build --@io_bazel_rules_scala//scala/settings:stamp_scala_import=False` in the `.bazelrc` file. Discussion in 
 https://github.com/bazelbuild/rules_scala/issues/1232 cleared that up. 

### Motivation

Reverting https://github.com/twitter-forks/rules_scala/pull/8 as it is not needed.